fixBugs in ConfirmRegister % Register
