+++ src/processes/auth/api/baseUrl.ts
@@ -1,0 +1,7 @@
+import axios from 'axios';
+
+const instance = axios.create({
+  baseURL: 'https://drake-singular-severely.ngrok-free.app/',
+});
+
+export default instance;
