+++ src/processes/auth/index.ts
@@ -1,4 +1,7 @@
 // export { default as getDeviceToken } from './api/getDeviceToken';
+export { default as instance } from './api/baseUrl';
+export { default as postRegisterUser } from './api/postRegisterUser';
+
 export { default as authReducer } from './model/slice';
 export { selectRole } from './model/selectors';
 export { setUserRole } from './model/operations';
