+++ src/processes/auth/model/operations.ts
@@ -1,28 +1,14 @@
 import { getDeviceToken } from '@/features/listening';
 import { createAction, createAsyncThunk } from '@reduxjs/toolkit';
-import axios from 'axios';
+import postRegisterUser from '../api/postRegisterUser';
 
 const setUserRole = createAction<'kid' | 'adult' | null>('auth/setUserRole');
 
 export { setUserRole };
 
-const instance = axios.create({
-  baseURL: 'https://drake-singular-severely.ngrok-free.app/',
-});
-
-export const registerUser = createAsyncThunk(
+export const postRegisterUserThank = createAsyncThunk(
   'auth/registerUser',
-    async (dataUser, thunkApi) => {
-    try {
-      const response = await instance.post('users/registration', dataUser);
-      return response.data;
-    } catch (error) {
-      alert(
-        `Oops! Something was wrong...`,
-      );
-      return thunkApi.rejectWithValue(error instanceof Error ? error.message : 'An error occurred');
-    }
-  },
+    postRegisterUser
 );
 
 export const getTokenThank = createAsyncThunk(
