+++ src/screens/adult/auth/resetpassword/ui/ConfirmRegisterScreen.tsx
@@ -10,8 +10,8 @@ import {
 import { styled } from 'nativewind';
 import { router } from 'expo-router';
 import { TouchableOpacity, Text } from 'react-native';
-import { selectUserData } from '@/processes/auth/model/selectors';
 import { useAppSelector } from '@/shared/lib';
+import { selectUserData } from '@/processes/auth/model/selectors';
 
 const Wrapper = styled(View);
 
@@ -37,8 +37,7 @@ const DigitInput = ({ digit, onChange }) => {
 
 const ConfirmRegisterScreen = () => {
   const [code, setCode] = useState(['', '', '', '']);
-  const userData = useAppSelector(selectUserData);
-  alert(`${userData?.message}`);
+ const regUserData = useAppSelector(selectUserData);
 
   const handleChange = (index, value) => {
     const newCode = [...code];
@@ -48,8 +47,9 @@ const ConfirmRegisterScreen = () => {
 
   const onSubmitForm = () => {
     const enteredCode = code.join('');
-    console.log('Entered Code:', enteredCode);
-    router.navigate('/auth/adult');
+    console.log('Entered Code & email:', enteredCode, regUserData?.dto.email);
+    setCode(['', '', '', '']);
+    router.navigate('/auth/adult/login');
   };
 
   const resendCode = () => {
@@ -94,10 +94,10 @@ const ConfirmRegisterScreen = () => {
         </Text>
       </TouchableOpacity>
       <PrimaryButton
-        text="Встановити новий пароль"
+        text="Підтвердити реєстрацію"
         onPress={onSubmitForm}
-        hint="Встановити новий пароль"
-        label="Встановити новий пароль"
+        hint="Підтвердити реєстрацію"
+        label="Підтвердити реєстрацію"
         role="button"
         classNames="w-full self-center mt-40"
       />
